BookStore/requirements.txt
marshmallow>=3
requests==2.26.0
flask
flask-restful
flask-migrate
flask-sqlalchemy
flask-marshmallow
marshmallow-sqlalchemy
python-dotenv
apispec[yaml]
apispec-webframeworks
tox
SQLAlchemy<1.4.0
mysqlclient

